Decouple handler from signature verification
